+++ src/test/java/ua/softserveinc/tc/paths/ChildDaoITPath.java
@@ -1,4 +1,4 @@
-package ua.softserveinc.tc.dao.paths;
+package ua.softserveinc.tc.paths;
 
 public final class ChildDaoITPath {
