+++ src/test/java/ua/softserveinc/tc/paths/TokenServiceITPath.java
@@ -1,0 +1,7 @@
+package ua.softserveinc.tc.paths;
+
+public final class TokenServiceITPath {
+
+    public static final String FIND_BY_EXISTING_USER = "classpath:tokenService/find-by-existing-user.xml";
+
+}
